refactor: update entity import
