+++ src/domains/tools/usecases/list-tools.usecase.spec.ts
@@ -2,7 +2,7 @@ import { afterEach, beforeEach, describe, expect, it } from "vitest";
 
 import { InMemoryToolRepository } from "@/domains/tools/repositories/implementations/in-memory";
 import { IToolRepository } from "@/domains/tools/repositories";
-import { ListToolsDTO } from "@/domains/tools";
+import { ListToolsDTO } from "@/domains/tools/entities";
 import { listToolUseCase, ListToolUseCase } from "@/domains/tools/usecases";
 
 describe("ListToolUseCase", () => {
